Fix problems with language specifications causing errors when submitting
There was an issue where languages are not using the correct name when submitting
to pamrel. This is causing an error when submitting, I have gone through every
language and comparing it to the value pamrel expects for those languages and
corrected those which were incorrect.

The issue referred is: https://notabug.org/Tsyesika/Pamrel/issues/1